sanitizer: Fix regexp matching of '-' in style attributes
Rework d6018f2539d271963c3e7f54f36ef11900363c69 and choose an alternative
approach to fixing bug 1623633.

Before that change, '-' could be matched either by itself, or by '\w-\w'. The
latter seemed to have the odd intent of avoiding leading or trailing '-' ...
but also meant it allowed 'a-aa-a' but not 'a-a-a'.

The primary intent with the change seemed to be to avoid a backtracking
explosion exploiding the ambiguous matching of '-'.

The chosen solution of allowing '\w-\w' but not '-' broke some tests and some
real world use cases.

There doesn't seem to be any risk from allowing leading or trailing or
consecutive '-' in styles. Thus, instead, this change will choose the path of
allowing '-' as a normal character and drop the special handling of '\w-\w'.